CODE/requirements.txt
langchain
langgraph
google-generativeai
pydantic
pinecone-client
streamlit
soundfile
numpy
deepgram-sdk